refactor: 랜덤숫자 생성기를 util 패키지로 옮기고 static으로 변환
