+++ src/main/java/RacingGameController.java
@@ -3,7 +3,6 @@ import domain.RacingGame;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import service.RandomNumberGenerator;
 import view.InputView;
 import view.OutputView;
 
@@ -16,7 +15,7 @@ public class RacingGameController {
     public RacingGameController() {
         this.inputView = new InputView();
         this.outputView = new OutputView();
-        this.racingGame = new RacingGame(new RandomNumberGenerator());
+        this.racingGame = new RacingGame();
     }
 
     public void run() {
