+++ src/main/java/util/NumberGenerator.java
@@ -1,4 +1,4 @@
-package service;
+package util;
 
 public interface NumberGenerator {
     int generate(int startInclusive, int endInclusive);
