+++ src/test/java/service/RacingGameTest.java
@@ -11,11 +11,12 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Queue;
 import org.junit.jupiter.api.Test;
+import util.NumberGenerator;
 
 class RacingGameTest {
     @Test
     void 이름들로_자동차들_생성_테스트() {
-        RacingGame racingGame = new RacingGame(new RandomNumberGenerator());
+        RacingGame racingGame = new RacingGame();
         List<String> carNames = Arrays.asList("aaa", "bbb", "ccc");
         racingGame.generateCars(carNames);
 
@@ -24,21 +25,21 @@ class RacingGameTest {
 
     @Test
     void 자동차_이름_중복_예외_테스트() {
-        RacingGame racingGame = new RacingGame(new RandomNumberGenerator());
+        RacingGame racingGame = new RacingGame();
         assertThatThrownBy(() -> racingGame.generateCars(Arrays.asList("bb", "bb", "cc")))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
     @Test
     void 자동차_개수_1개_일_때_예외_테스트() {
-        RacingGame racingGame = new RacingGame(new RandomNumberGenerator());
+        RacingGame racingGame = new RacingGame();
         assertThatThrownBy(() -> racingGame.generateCars(Arrays.asList("bb")))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
     @Test
     void 값에_따른_자동차_진행_테스트() {
-        RacingGame racingGame = new RacingGame(new TestNumberGenerator());
+        RacingGame racingGame = new RacingGame();
         List<String> carNames = new ArrayList<>(List.of("aaa", "bbb", "ccc"));
         racingGame.generateCars(carNames);
 
@@ -52,7 +53,7 @@ class RacingGameTest {
 
     @Test
     void 게임_승자_테스트() {
-        RacingGame racingGame = new RacingGame(new TestNumberGenerator());
+        RacingGame racingGame = new RacingGame();
         List<String> carNames = new ArrayList<>(List.of("aaa", "bbb", "ccc"));
         racingGame.generateCars(carNames);
